[MINOR] Use CommonUtils instead of U in binary
