+++ modules/binary/api/src/main/java/org/apache/ignite/marshaller/Marshallers.java
@@ -17,10 +17,8 @@
 
 package org.apache.ignite.marshaller;
 
-import java.io.OutputStream;
 import java.io.Serializable;
 import java.util.Iterator;
-import org.apache.ignite.IgniteCheckedException;
 import org.apache.ignite.IgniteCommonsSystemProperties;
 import org.apache.ignite.internal.marshaller.optimized.OptimizedMarshaller;
 import org.apache.ignite.internal.util.CommonUtils;
@@ -83,54 +81,5 @@ public class Marshallers {
      */
     public static OptimizedMarshaller optimized(boolean requireSer) {
         return factory.optimized(requireSer);
-    }
-
-    /**
-     * Marshals object to byte array.
-     * <p/>
-     * This method wraps marshaller invocations and guaranty throws {@link IgniteCheckedException} in fail case.
-     *
-     * @param marsh Marshaller.
-     * @param obj Object to marshal.
-     * @return Byte array.
-     * @throws IgniteCheckedException If marshalling failed.
-     */
-    public static byte[] marshal(Marshaller marsh, Object obj) throws IgniteCheckedException {
-        assert marsh != null;
-
-        try {
-            return marsh.marshal(obj);
-        }
-        catch (IgniteCheckedException e) {
-            throw e;
-        }
-        catch (Exception e) {
-            throw new IgniteCheckedException(e);
-        }
-    }
-
-    /**
-     * Marshals object to byte array.
-     * <p/>
-     * This method wraps marshaller invocations and guaranty throws {@link IgniteCheckedException} in fail case.
-     *
-     * @param marsh Marshaller.
-     * @param obj Object to marshal.
-     * @param out Output stream.
-     * @throws IgniteCheckedException If marshalling failed.
-     */
-    public static void marshal(Marshaller marsh, @Nullable Object obj, OutputStream out)
-        throws IgniteCheckedException {
-        assert marsh != null;
-
-        try {
-            marsh.marshal(obj, out);
-        }
-        catch (IgniteCheckedException e) {
-            throw e;
-        }
-        catch (Exception e) {
-            throw new IgniteCheckedException(e);
-        }
     }
 }
