+++ modules/core/src/main/java/org/apache/ignite/internal/IgniteCodeGeneratingFail.java
