+++ modules/core/src/main/java/org/apache/ignite/internal/binary/BinaryReaderExImpl.java
@@ -35,9 +35,10 @@ import org.apache.ignite.binary.BinaryObject;
 import org.apache.ignite.binary.BinaryObjectException;
 import org.apache.ignite.binary.BinaryRawReader;
 import org.apache.ignite.internal.binary.streams.BinaryInputStream;
-import org.apache.ignite.internal.util.CommonUtils;
+import org.apache.ignite.internal.util.IgniteUtils;
 import org.apache.ignite.internal.util.typedef.internal.S;
 import org.apache.ignite.internal.util.typedef.internal.SB;
+import org.apache.ignite.internal.util.typedef.internal.U;
 import org.apache.ignite.marshaller.Marshallers;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -1702,13 +1703,13 @@ class BinaryReaderExImpl implements BinaryReaderEx {
     /** {@inheritDoc} */
     @Override public @Nullable Object deserialize() throws BinaryObjectException {
         String newName = ctx.igniteInstanceName();
-        String oldName = CommonUtils.setCurrentIgniteName(newName);
+        String oldName = IgniteUtils.setCurrentIgniteName(newName);
 
         try {
             return deserialize0();
         }
         finally {
-            CommonUtils.restoreOldIgniteName(oldName, newName);
+            IgniteUtils.restoreOldIgniteName(oldName, newName);
         }
     }
 
@@ -1999,7 +2000,7 @@ class BinaryReaderExImpl implements BinaryReaderEx {
                     throw new BinaryObjectException("Cannot find metadata for object with compact footer " +
                         "(Ignite work directory might have been cleared after restart. Make sure that IGNITE_HOME " +
                         "does not point to a temp folder or any other folder that is destroyed/cleared on restarts) [" +
-                        "typeId=" + typeId + ", IGNITE_HOME='" + CommonUtils.getIgniteHome() + "']");
+                        "typeId=" + typeId + ", IGNITE_HOME='" + U.getIgniteHome() + "']");
 
                 Collection<BinarySchema> existingSchemas = meta.schemas();
