+++ modules/core/src/main/java/org/apache/ignite/internal/binary/BinaryWriterExImpl.java
@@ -34,9 +34,9 @@ import org.apache.ignite.binary.BinaryObjectException;
 import org.apache.ignite.binary.BinaryRawWriter;
 import org.apache.ignite.internal.UnregisteredClassException;
 import org.apache.ignite.internal.binary.streams.BinaryOutputStream;
-import org.apache.ignite.internal.util.CommonUtils;
+import org.apache.ignite.internal.util.IgniteUtils;
 import org.apache.ignite.internal.util.typedef.internal.A;
-import org.apache.ignite.marshaller.Marshallers;
+import org.apache.ignite.internal.util.typedef.internal.U;
 import org.jetbrains.annotations.Nullable;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
@@ -131,13 +131,13 @@ class BinaryWriterExImpl implements BinaryWriterEx {
      */
     void marshal(Object obj, boolean enableReplace) throws BinaryObjectException {
         String newName = ctx.igniteInstanceName();
-        String oldName = CommonUtils.setCurrentIgniteName(newName);
+        String oldName = IgniteUtils.setCurrentIgniteName(newName);
 
         try {
             marshal0(obj, enableReplace);
         }
         finally {
-            CommonUtils.restoreOldIgniteName(oldName, newName);
+            IgniteUtils.restoreOldIgniteName(oldName, newName);
         }
     }
 
@@ -174,7 +174,7 @@ class BinaryWriterExImpl implements BinaryWriterEx {
             out.writeByte(GridBinaryMarshaller.OPTM_MARSH);
 
             try {
-                byte[] arr = Marshallers.marshal(ctx.optimizedMarsh(), obj);
+                byte[] arr = U.marshal(ctx.optimizedMarsh(), obj);
 
                 writeInt(arr.length);
